requirements.txt
Flask==2.2.2
flask-sqlalchemy==3.0.2
Flask-WTF
email-validator
Werkzeug>2.1.1
flask-bcrypt
mysqlclient
flask-login==0.6.2
openpyxl==3.0.10
flask-user 

Flask-Session==0.4.0
pycryptodome==3.16.0
itsdangerous==2.1.2
